// server/src/models/enums.ts
export enum UserStatus {
  ONLINE = 'ONLINE',
  IN_GAME = 'IN_GAME',
  OFFLINE = 'OFFLINE',
}

export enum RoomStatus {
  WAITING = 'WAITING',
  IN_PROGRESS = 'IN_PROGRESS',
  FINISHED = 'FINISHED',
}

export enum GameStatus {
  RUNNING = 'RUNNING',
  FINISHED = 'FINISHED',
  DRAW = 'DRAW',
}

export enum MoveSymbol {
  X = 'X',
  O = 'O',
}

export enum ChatMessageType {
  USER = 'USER',
  SYSTEM = 'SYSTEM',
}